refactor requests / responses
