+++ src/api/game_api.go
@@ -2,78 +2,19 @@ package api
 
 import (
 	"encoding/json"
-	"fmt"
 	"golife/internal"
 	"golife/server"
 	"log"
 	"net/http"
 	"path/filepath"
-	"strconv"
-	"strings"
 )
-
-const (
-	MAX_MATRIX_ROWS = 100
-	MAX_MATRIX_COLS = 100
-)
-
-// extract request / responses to another file in the same package
-type ApiErrorRes struct {
-	Code int
-	Msg  string
-}
-
-type RenderMatrixReq struct {
-	template string
-	offsetX  int
-	offsetY  int
-	rows     int
-	cols     int
-}
-
-type RenderMatrixRes struct {
-	Matrix [][]uint8
-}
-
-func (res *RenderMatrixRes) MarshalJson() ([]byte, error) {
-	var matrix string
-	if res.Matrix == nil {
-		matrix = "null"
-	} else {
-		matrix = strings.Join(strings.Fields(fmt.Sprintf("%d", res.Matrix)), ",")
-	}
-	jsonResult := fmt.Sprintf(`{"matrix":%s}`, matrix)
-	return []byte(jsonResult), nil
-}
 
 func RenderMatrix(w http.ResponseWriter, req *http.Request) {
 	var matrix [][]uint8
 	env := server.GetEnv()
+	request := new(RenderMatrixReq).createFromQueryUrl(req.URL.Query())
 
-	// refactor me
-	req_query := req.URL.Query()
-	request := RenderMatrixReq{
-		template: req_query.Get("template"),
-	}
-	if req_query.Has("offsetX") {
-		request.offsetX, _ = strconv.Atoi(req_query.Get("offsetX"))
-	}
-	if req_query.Has("offsetY") {
-		request.offsetY, _ = strconv.Atoi(req_query.Get("offsetY"))
-	}
-	if req_query.Has("rows") {
-		request.rows, _ = strconv.Atoi(req_query.Get("rows"))
-		if request.rows > MAX_MATRIX_ROWS {
-			request.rows = MAX_MATRIX_ROWS
-		}
-	}
-	if req_query.Has("cols") {
-		request.cols, _ = strconv.Atoi(req_query.Get("cols"))
-		if request.cols > MAX_MATRIX_COLS {
-			request.cols = MAX_MATRIX_COLS
-		}
-	}
-
+	log.Printf("render req: %+v", request)
 	if request.template != "" {
 		template_path := filepath.Join(env.Tempalate_folder, request.template)
 		matrix = internal.LoadFieldMatrix(template_path, request.offsetX, request.offsetY, request.rows, request.cols)
@@ -91,10 +32,6 @@ func RenderMatrix(w http.ResponseWriter, req *http.Request) {
 	}
 	w.WriteHeader(http.StatusOK)
 	w.Write(json_response)
-}
-
-type GetNextStepReq struct {
-	Matrix [][]uint8 `json:"matrix"`
 }
 
 func GetNextStep(w http.ResponseWriter, req *http.Request) {
